% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covshift.R
\name{covshift}
\alias{covshift}
\title{Risk estimation and inference under covariate shift}
\usage{
covshift(
  learners,
  loss.fun,
  A,
  X,
  Y,
  nfolds = 5,
  cond.loss.SL.control = list(family = gaussian(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  prop.score.SL.control = list(family = binomial(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  conf.level = 0.95,
  losses = NULL
)
}
\arguments{
\item{learners}{a list of learners (functions) that can take covariate \code{X} as input; can also be a function if only one learner is of interest}

\item{loss.fun}{a user specified loss function depending on covariate \code{X}, outcome \code{Y} and the learner that defines the risk (expected loss in the target population for the learner). Must take in three arguments \code{X}, \code{Y} and \code{learner} (an element of \code{learners}) and output a numeric vector with length equal to the total sample size}

\item{A}{a vector of binary indicator of data point coming from source population (encoded by \code{1}) or target population (encoded by \code{0})}

\item{X}{data frame of covariates}

\item{Y}{data frame of outcomes}

\item{nfolds}{number of folds used for the cross-fit estimator; default to 5}

\item{cond.loss.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the loss conditional on covariate \code{X} (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=gaussian(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{prop.score.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the propensity score function (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=binomial(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{conf.level}{confidence level in the output (asymptotic) confidence interval}

\item{losses}{alternative input in place of \code{learners}, \code{loss.fun} and \code{Y}; a matrix of loss function evaluated at each data point and for each learner (row index is for data point; column index is for learner); default to \code{NULL}}
}
\value{
a list with the following elements:
\itemize{
\item \code{estimates}: a data frame with each row being the output for one learner in \code{learners}
\item \code{empirical.IF}: a matrix of influence functions evaluated at empirical data poins. Each column corresponds to a learner in \code{learners} and each row corresponds to a data point
}
}
\description{
Using fully observed data from both source and target populations, estimate the risk of a list of learners in the target population under covariate shift assumption (that is, the distributions of outcome given covariate are identical in both populations but covariate distributions might differ). The estimator is semiparametrically efficient when both nuisance functions are consistently estimated and is still consistent if one is inconsistently estimated.
}
\section{Using \code{empirical.IF} to infer contrasts of risks}{

The output \code{empirical.IF} can be used to infer arbitrary smooth functions of the risks of several learners, e.g., contrasts, by using the Delta method for influence functions. Here, differences and ratios are illustrated as examples.

Suppose that the followings are computed:
\itemize{
\item point estimates \code{r1} and \code{r2} of risks corresponding to two learners
\item their respective influence functions evaluated at empirical data being (vectors) \code{IF1} and \code{IF2}
}

To infer the difference between the two risks, one can use \code{r1-r2} as the point estimate and \code{IF1-IF2} as the corresponding influence function, so that the standard error should be \code{sqrt(mean((IF1-IF2)^2)/leangth(IF1))}. One can construct confidence intervals based on normal approximation.

To infer the ratio between the two risks (assuming that both risks are positive), one can use \code{r1/r2} as the point estimate and \code{IF1/r2-IF2*r1/r2^2} as the corresponding influence function, so that the standard error should be \code{sqrt(mean((IF1/r2-IF2*r1/r2^2)^2)/leangth(IF1))}. One can construct confidence intervals based on normal approximation.
}

