% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_conshift.R
\name{cross_entropy_feat_conshift}
\alias{cross_entropy_feat_conshift}
\title{Cross-entropy risk estimation and inference under concept shift in feature/covariate}
\usage{
cross_entropy_feat_conshift(
  learners,
  A,
  X,
  Y,
  nfolds = 5,
  cond.Y.SL.control = list(family = binomial(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  conf.level = 0.95,
  Yhat = NULL
)
}
\arguments{
\item{learners}{a list of learners (functions) that can take covariate \code{X} as input; can also be a function if only one learner is of interest}

\item{A}{a vector of binary indicator of data point coming from source population (encoded by \code{1}) or target population (encoded by \code{0})}

\item{X}{data frame of features/covariates}

\item{Y}{data frame of binary labels/outcomes (0 or 1) with one column}

\item{nfolds}{number of folds used for the cross-fit estimator; default to 5}

\item{cond.Y.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the conditional mean of label/outcome \code{Y} given feature/covariate \code{X} (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=binomial(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{conf.level}{confidence level in the output (asymptotic) confidence interval}

\item{Yhat}{alternative input in place of \code{learners}; a matrix of predicted P(Y=1) at each data point and for each learner (row index is for data point; column index is for learner); default to \code{NULL}}
}
\value{
a list with the following elements:
\itemize{
\item \code{estimates}: a data frame with each row being the output for one learner in \code{learners}
\item \code{empirical.IF}: a matrix of influence functions evaluated at empirical data poins. Each column corresponds to a learner in \code{learners} and each row corresponds to a data point
}
}
\description{
Using data from both source population (outcome may be missing) and target population (fully observed), estimate the cross-entropy risk of a list of learners in the target population under concept shift (in feature/covariate) assumption (that is, feature/covariate distributions are identical in both populations but the distributions of label/outcome given feature/covariate might differ). The estimator is semiparametrically efficient when the nuisance function is consistently estimated and is asymptotically normal even if the nuisance function is inconsistently estimated. This function is more computationally efficient than using the more general function \code{\link{feat_conshift}} when multiple learners are of interest because the structure of the loss function is exploited.
}
\section{Using \code{empirical.IF} to infer contrasts of risks}{

See \link{covshift}
}

