% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labshift.R
\name{labshift}
\alias{labshift}
\title{Risk estimation and inference under label shift}
\usage{
labshift(
  learners,
  loss.fun,
  A,
  X,
  Y,
  nfolds = 5,
  cond.loss.SL.control = list(family = gaussian(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  prop.score.SL.control = list(family = binomial(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  conf.level = 0.95,
  losses = NULL
)
}
\arguments{
\item{learners}{a list of learners (functions) that can take covariate \code{X} as input; can also be a function if only one learner is of interest}

\item{loss.fun}{a user specified loss function depending on covariate \code{X}, outcome \code{Y} and the learner that defines the risk (expected loss in the target population for the learner). Must take in three arguments \code{X}, \code{Y} and \code{learner} (an element of \code{learners}) and output a numeric vector with length equal to the total sample size}

\item{A}{a vector of binary indicator of data point coming from source population (encoded by \code{1}) or target population (encoded by \code{0})}

\item{X}{data frame of covariates}

\item{Y}{data frame of outcomes}

\item{nfolds}{number of folds used for the cross-fit estimator; default to 5}

\item{cond.loss.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the loss conditional on outcome \code{Y} (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=gaussian(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{prop.score.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the propensity score function (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=binomial(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{conf.level}{confidence level in the output (asymptotic) confidence interval}

\item{losses}{alternative input in place of \code{learners}, \code{loss.fun} and \code{X}; a matrix of loss function evaluated at each data point and for each learner (row index is for data point; column index is for learner); default to \code{NULL}}
}
\value{
a list with the following elements:
\itemize{
\item \code{estimates}: a data frame with each row being the output for one learner in \code{learners}
\item \code{empirical.IF}: a matrix of influence functions evaluated at empirical data poins. Each column corresponds to a learner in \code{learners} and each row corresponds to a data point
}
}
\description{
Using fully observed data from both source and target populations, estimate the risk of a list of learners in the target population under label shift assumption (that is, the distributions of covariate given outcome are identical in both populations but outcome distributions might differ). The estimator is semiparametrically efficient when both nuisance functions are consistently estimated and is asymptotically normal even if both are inconsistently estimated.
}
\section{Using \code{empirical.IF} to infer contrasts of risks}{

See \link{covshift}
}

