% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covshift.R
\name{linMSE_covshift}
\alias{linMSE_covshift}
\title{Linear MSE estimation and inference under covariate shift}
\usage{
linMSE_covshift(
  learners,
  A,
  X,
  Y,
  nfolds = 5,
  cond.Y.SL.control = list(family = gaussian(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  prop.score.SL.control = list(family = binomial(), SL.library = c("SL.glm", "SL.gam",
    "SL.randomForest")),
  conf.level = 0.95,
  Yhat = NULL
)
}
\arguments{
\item{learners}{a list of learners (functions) that can take covariate \code{X} as input and outputs a numeric vector with length being \code{nrow(X)}; can also be a function if only one learner is of interest}

\item{A}{a vector of binary indicator of data point coming from source population (encoded by \code{1}) or target population (encoded by \code{0})}

\item{X}{data frame of covariates}

\item{Y}{data frame of outcomes with one column}

\item{nfolds}{number of folds used for the cross-fit estimator; default to 5}

\item{cond.Y.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the conditional mean of outcome \code{Y} given covariate \code{X} (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=gaussian(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{prop.score.SL.control}{a named list containing options passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to estimate the propensity score function (a nuisance function). Must not specify \code{Y}, \code{X} or \code{newX}. Default to \code{list(family=binomial(),SL.library=c("SL.glm","SL.gam","SL.randomForest"))}}

\item{conf.level}{confidence level in the output (asymptotic) confidence interval}

\item{Yhat}{alternative input in place of \code{learners}; a matrix of predicted P(Y=1) at each data point and for each learner (row index is for data point; column index is for learner); default to \code{NULL}}
}
\value{
a list with the following elements:
\itemize{
\item \code{estimates}: a data frame with each row being the output for one learner in \code{learners}
\item \code{empirical.IF}: a matrix of influence functions evaluated at empirical data poins. Each column corresponds to a learner in \code{learners} and each row corresponds to a data point
}
}
\description{
Using fully observed data from both source and target populations, estimate the linear MSE of a list of learners in the target population under covariate shift assumption (that is, the distributions of outcome given covariate are identical in both populations but covariate distributions might differ). The estimator is semiparametrically efficient when both nuisance functions are consistently estimated and is still consistent if one is inconsistently estimated. This function is more computationally efficient than using the more general function \code{\link{covshift}} when multiple learners are of interest because the structure of the loss function is exploited.
}
\section{Using \code{empirical.IF} to infer contrasts of risks}{

See \link{covshift}
}

