% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{RiskEstDShift-package}
\alias{RiskEstDShift-package}
\title{RiskEstDShift package}
\description{
In this package, we implement efficient and multiply robust cross-fit target population risk estimators under one of the following four dataset shift conditions:
\itemize{
\item full-data covariate shift: The distribution of \eqn{Y|X}{Y|X} is same in source and target populations and \eqn{(X,Y)}{(X,Y)} is observed in data from both populations
\item full-data label shift: The distribution of \eqn{X|Y}{X|Y} is same in source and target populations and \eqn{(X,Y)}{(X,Y)} is observed in data from both populations
\item concept shift in the features: The distribution of \eqn{X}{X} is same in source and target populations
\item concept shift in the labels: The distribution of \eqn{Y}{Y} is same in source and target populations
}
}
